ticket all and users with images

+++ ComsiteDesk.ERP.DB.Core/GeneralModels/ModelBase.cs
@@ -11,9 +11,9 @@ namespace ComsiteDesk.ERP.DB.Core.Models
             IsActive = true;
         }
         public DateTime DateCreated { get; set; }
-        public int CreatedBy { get; set; }
+        public long CreatedBy { get; set; }
         public DateTime? DateModified { get; set; }
-        public int? ModifiedBy { get; set; }       
+        public long? ModifiedBy { get; set; }       
         public bool IsActive { get; set; }
     }
 }
